requirements.txt
flask==3.0.0
twilio==8.10.0
openai==1.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
langdetect==1.0.9
APScheduler==3.10.4
gunicorn==21.2.0